+++ src/client/widgets/common/widget.js
@@ -201,6 +201,7 @@ class Widget extends EventEmitter {
         this.isValueChanging = false
         this.valueChangingTimeout = null
         this.valueChangingQueue = {}
+        this.value = options.value
 
         this.parsers = {}
         this.parsersLocalScope = options.locals || {}