+++ src/client/widgets/pads/xy.js
@@ -84,8 +84,6 @@ module.exports = class Xy extends Pad {
             e.stopPropagation = true
         })
 
-        this.value = []
-
         touchstate(this, {element: this.canvas})
 
         this.active = false