+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 - bug fixes
     - code editor: issue when middle clicking in the editor on linux
     - code editor: modern js operators such as `**` not recognized
-    - sliders: prevent error when `steps` is not a number, add `auto` mode for using ranges values as steps (it did not work before) 
+    - sliders: prevent error when `steps` is not a number, add `auto` mode for using ranges values as steps (it did not work before)
+    - pads, range: value not initialized as an array, causing errors in some cases
 
 ## 1.25.4